Fix an integer overflow issue.

+++ src/main.cpp
@@ -103,7 +103,7 @@ void setup() {
     if (FILE_LOGGING) {
         log("Creating log file... ", false);
         char log_file_name[] = "log_file_000.txt";
-        for (uint8_t i = 0; i < MAX_LOG_FILES; i++) {
+        for (uint16_t i = 0; i < MAX_LOG_FILES; i++) {
             // Increment until we can find a log file slot.
 
             // Need to add 32 to get ASCII number characters.
@@ -136,7 +136,7 @@ void setup() {
     // SET UP DATA FILE
     log("Creating data file...");
     char data_file_name[] = "data_file_000.csv";
-    for (uint8_t i = 0; i < MAX_DATA_FILES; i++) {
+    for (uint16_t i = 0; i < MAX_DATA_FILES; i++) {
         // Increment until we can find a data file slot.
 
         // Need to add 32 to get ASCII number characters.
